+++ src/components/logo/index.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import * as React from "react";
-import { useRouter } from "next/navigation";
+import { useRouter } from "next-intl/client";
 import { cn } from "@/lib";
 
 interface LogoProps {